Moved all stats logic to the backend.  Added user check to the grabbed stats and backported that check to Edit Stat Xp Check
