+++ ExpressedRealms.Repositories.Characters/Stats/CharacterStatRepository.cs
@@ -91,7 +91,7 @@ internal sealed class CharacterStatRepository(
             return Result.Fail(new FluentValidationFailure(result.ToDictionary()));
         
         var character = await context
-            .Characters.Where(x => x.Id == dto.CharacterId)
+            .Characters.Where(x => x.Id == dto.CharacterId && x.Player.UserId == userContext.CurrentUserId())
             .Include(x => x.AgilityStatLevel)
             .Include(x => x.StrengthStatLevel)
             .Include(x => x.ConstitutionStatLevel)
@@ -175,5 +175,82 @@ internal sealed class CharacterStatRepository(
         }
 
         return Result.Ok();
+    }
+
+    public async Task<Result<List<SmallStatInfo>>> GetAllStats(int characterId)
+    {
+        var character = await context
+            .Characters.Include(x => x.AgilityStatLevel)
+            .Include(x => x.ConstitutionStatLevel)
+            .Include(x => x.DexterityStatLevel)
+            .Include(x => x.StrengthStatLevel)
+            .Include(x => x.IntelligenceStatLevel)
+            .Include(x => x.WillpowerStatLevel)
+            .FirstOrDefaultAsync(x => x.Id == characterId && x.Player.UserId == userContext.CurrentUserId(), cancellationToken);
+
+        if (character is null)
+            return Result.Fail(new NotFoundFailure("Character"));
+
+        var statTypes = await context.StateTypes.ToListAsync(cancellationToken);
+        
+        var characterStats = new List<SmallStatInfo>()
+        {
+            new()
+            {
+                StatTypeId = StatType.Agility,
+                Bonus = character.AgilityStatLevel.Bonus,
+                Level = character.AgilityStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Agility)
+                    .ShortName
+            },
+            new()
+            {
+                StatTypeId = StatType.Constitution,
+                Bonus = character.ConstitutionStatLevel.Bonus,
+                Level = character.ConstitutionStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Constitution)
+                    .ShortName
+            },
+            new()
+            {
+                StatTypeId = StatType.Dexterity,
+                Bonus = character.DexterityStatLevel.Bonus,
+                Level = character.DexterityStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Dexterity)
+                    .ShortName
+            },
+            new()
+            {
+                StatTypeId = StatType.Strength,
+                Bonus = character.StrengthStatLevel.Bonus,
+                Level = character.StrengthStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Strength)
+                    .ShortName
+            },
+            new()
+            {
+                StatTypeId = StatType.Intelligence,
+                Bonus = character.IntelligenceStatLevel.Bonus,
+                Level = character.IntelligenceStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Intelligence)
+                    .ShortName
+            },
+            new()
+            {
+                StatTypeId = StatType.Willpower,
+                Bonus = character.WillpowerStatLevel.Bonus,
+                Level = character.WillpowerStatLevel.Id,
+                ShortName = statTypes
+                    .First(x => x.Id == (byte)StatType.Willpower)
+                    .ShortName
+            }
+        };
+
+        return Result.Ok(characterStats);
     }
 }
