+++ ExpressedRealms.Repositories.Characters/Stats/ICharacterStatRepository.cs
@@ -7,4 +7,5 @@ public interface ICharacterStatRepository
 {
     Task<Result<SingleStatInfo>> GetDetailedStatInfo(GetDetailedStatInfoDto dto);
     Task<Result> UpdateCharacterStat(EditStatDto dto);
+    Task<Result<List<SmallStatInfo>>> GetAllStats(int characterId);
 }
