+++ ExpressedRealms.Server/EndPoints/CharacterEndPoints/StatDTOs/SmallStatInfo.cs
@@ -2,17 +2,17 @@ using ExpressedRealms.Repositories.Characters.Stats.Enums;
 
 namespace ExpressedRealms.Server.EndPoints.CharacterEndPoints.StatDTOs;
 
-public class SmallStatInfo
+public class SmallStatInfo(Repositories.Characters.Stats.DTOs.SmallStatInfo smallStatInfo)
 {
     /// <example>WIL</example>
-    public string ShortName { get; set; }
+    public string ShortName { get; set; } = smallStatInfo.ShortName;
 
     /// <example>1</example>
-    public int Level { get; set; }
+    public int Level { get; set; } = smallStatInfo.Level;
 
     /// <example>-1</example>
-    public int Bonus { get; set; }
+    public int Bonus { get; set; } = smallStatInfo.Bonus;
 
     /// <example>6</example>
-    public StatType StatTypeId { get; set; }
+    public StatType StatTypeId { get; set; } = smallStatInfo.StatTypeId;
 }
